+++ src/targets/git.ts
@@ -34,7 +34,7 @@ export function makeGit(): Target {
 			ruleCompile({
 				compiled: null,
 				excludes: true,
-				pattern: [".git", ".DS_Store"],
+				list: [".git", ".DS_Store"],
 			}),
 		],
 	}