+++ src/targets/jsr.ts
@@ -28,7 +28,7 @@ export function makeJSR(): Target {
 		ruleCompile({
 			compiled: null,
 			excludes: true,
-			pattern: [".git", ".DS_Store"],
+			list: [".git", ".DS_Store"],
 		}),
 	]