+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Rename `Rule.pattern`, `PatternCache.patternContext` to `list`.
 - Fix `match.ignored`. Were always false.
 - `invert` now supports the value `2`, which enables all paths.
 - `Source.rules` is now an array of rules.